django_final/requirement.txt
asgiref==3.8.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.0.0
chardet==3.0.4
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.0
Django==5.0.3
djangorestframework==3.15.1
fonttools==4.53.1
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.2
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
paho-mqtt==2.0.0
pandas==2.2.2
pillow==10.3.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
rfc3986==1.5.0
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.1.0
setuptools==71.1.0
shell==1.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
threadpoolctl==3.5.0
tqdm==4.66.5
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
whitenoise==6.7.0
zope.interface==6.2
